moved request handler to a seperate file
